refactor: move handleDagPb to plugin
