+++ packages/verified-fetch/src/plugins/plugin-handle-dag-cbor.ts
@@ -19,7 +19,7 @@ export class DagCborPlugin implements FetchHandlerPlugin {
   async handle (context: PluginContext, pluginOptions: PluginOptions): Promise<Response> {
     const { cid, path, resource, accept } = context
     const { logger, options, getBlockstore, handleServerTiming, withServerTiming = false } = pluginOptions
-    const log = logger.forComponent('json-plugin')
+    const log = logger.forComponent('dag-cbor-plugin')
     const session = options?.session ?? true
 
     log.trace('fetching %c/%s', cid, path)
@@ -84,68 +84,3 @@ export class DagCborPlugin implements FetchHandlerPlugin {
     return response
   }
 }
-
-// private async handleDagCbor ({ resource, cid, path, accept, session, options, withServerTiming }: FetchHandlerFunctionArg): Promise<Response> {
-//   this.log.trace('fetching %c/%s', cid, path)
-//   let terminalElement: ObjectNode
-//   const blockstore = this.getBlockstore(cid, resource, session, options)
-
-//   // need to walk path, if it exists, to get the terminal element
-//   const pathDetails = await this.handleServerTiming('path-walking', '', async () => handlePathWalking({ cid, path, resource, options, blockstore, log: this.log, withServerTiming }), withServerTiming)
-
-//   if (pathDetails instanceof Response) {
-//     return pathDetails
-//   }
-//   const ipfsRoots = pathDetails.ipfsRoots
-//   if (isObjectNode(pathDetails.terminalElement)) {
-//     terminalElement = pathDetails.terminalElement
-//   } else {
-//     // this should never happen, but if it does, we should log it and return notSupportedResponse
-//     this.log.error('terminal element is not a dag-cbor node')
-//     return notSupportedResponse(resource, 'Terminal element is not a dag-cbor node')
-//   }
-
-//   const block = terminalElement.node
-
-//   let body: string | Uint8Array
-
-//   if (accept === 'application/octet-stream' || accept === 'application/vnd.ipld.dag-cbor' || accept === 'application/cbor') {
-//     // skip decoding
-//     body = block
-//   } else if (accept === 'application/vnd.ipld.dag-json') {
-//     try {
-//       // if vnd.ipld.dag-json has been specified, convert to the format - note
-//       // that this supports more data types than regular JSON, the content-type
-//       // response header is set so the user knows to process it differently
-//       const obj = ipldDagCbor.decode(block)
-//       body = ipldDagJson.encode(obj)
-//     } catch (err) {
-//       this.log.error('could not transform %c to application/vnd.ipld.dag-json', err)
-//       return notAcceptableResponse(resource)
-//     }
-//   } else {
-//     try {
-//       body = dagCborToSafeJSON(block)
-//     } catch (err) {
-//       if (accept === 'application/json') {
-//         this.log('could not decode DAG-CBOR as JSON-safe, but the client sent "Accept: application/json"', err)
-
-//         return notAcceptableResponse(resource)
-//       }
-
-//       this.log('could not decode DAG-CBOR as JSON-safe, falling back to `application/octet-stream`', err)
-//       body = block
-//     }
-//   }
-
-//   const response = okResponse(resource, body)
-
-//   if (accept == null) {
-//     accept = body instanceof Uint8Array ? 'application/octet-stream' : 'application/json'
-//   }
-
-//   response.headers.set('content-type', accept)
-//   setIpfsRoots(response, ipfsRoots)
-
-//   return response
-// }
